requirements.txt
pyside6
pyserial
requests
python-dotenv
firebase_admin
pycryptodome